Add files via upload
The container code, which is allocated when containers are created, takes the following format. 
The current year value between 2000 and 2099 (both included) 
 Forward slash (/) 
 The current month value between 01 and 12 (both included) 
 Forward slash (/) 
 A ‘B’ or a ‘C’, depending on the container type. 
A serial number starting from 1, running up to 9999. Each new container is given the next serial number. 

When a user creates a box or cylinder, the objects should be added to an unallocated container 
list. The container codes of these items should be displayed on the user interface using some 
model/view convenience class instance. The user should be able to backup this list of unallocated 
containers (remembering that you cannot simply back up the pointers to the container objects as 
they may be deleted before the restore is done); when restoring the unallocated container list, the 
container objects should be recreated and added to the list, and the model/view convenience 
object appropriately updated. Use an appropriate design pattern for this backup/restore functionality